refactor: turn tuono-fs-router-vite-plugin into tuono-react-vite-plugin
