+++ packages/tuono-react-vite-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tuono-fs-router-vite-plugin",
+  "name": "tuono-react-vite-plugin",
   "version": "0.19.4",
   "description": "Plugin for the tuono's file system router. Tuono is the react/rust fullstack framework",
   "homepage": "https://tuono.dev",
@@ -16,7 +16,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tuono-labs/tuono.git",
-    "directory": "packages/tuono-fs-router-vite-plugin"
+    "directory": "packages/tuono-react-vite-plugin"
   },
   "keywords": [],
   "author": "Valerio Ageno",
