+++ packages/tuono-react-vite-plugin/src/fs-routing/sort-route-nodes.ts
@@ -1,4 +1,5 @@
-import type { RouteNode } from './types'
+import type { RouteNode } from '../types'
+
 import { multiSortBy } from './utils'
 import { LAYOUT_PATH_ID } from './constants'
