+++ packages/tuono-react-vite-plugin/src/index.ts
@@ -1,0 +1,1 @@
+export { TuonoReactPlugin } from './plugin'
