+++ packages/tuono-react-vite-plugin/tests/generator/catch_all/routeTree.expected.ts
