+++ packages/tuono-react-vite-plugin/tests/generator/catch_all/routes/index.tsx
