+++ packages/tuono-react-vite-plugin/tests/generator/catch_all/routes/posts/[...catch_all].tsx
