+++ packages/tuono-react-vite-plugin/tests/generator/mdx/routes/__layout.tsx
