+++ packages/tuono-react-vite-plugin/tests/generator/multi-level-root-dynamic/routeTree.expected.ts
