+++ packages/tuono-react-vite-plugin/tests/generator/multi-level-root-dynamic/routes/posts/my-post.tsx
