+++ packages/tuono-react-vite-plugin/tests/generator/multi-level/routeTree.expected.ts
