+++ packages/tuono-react-vite-plugin/tests/generator/multi-level/routes/posts/my-post.tsx
