+++ packages/tuono-react-vite-plugin/tests/generator/single-level/routeTree.expected.ts
