+++ packages/tuono-react-vite-plugin/tests/generator/single-level/routes/__layout.tsx
