+++ packages/tuono-react-vite-plugin/tests/generator/single-level/routes/about.tsx
