+++ packages/tuono-react-vite-plugin/tests/generator/single-level/routes/index.tsx
